Update cpu_mask type to uint64_t in configuration adapter
The default processor affinity mask was computed and stored as uint32_t,
which is inconsistent with the uint64_t cpu_mask_ field in OsalConfig
and the setaffinity(uint64_t) API. Widen the return type and remove
the narrowing cast so the adapter correctly passes 64-bit masks through.

+++ score/launch_manager/src/daemon/src/configuration/configuration_adapter.cpp
@@ -28,8 +28,8 @@ namespace {
 constexpr const char* kAliveInterfaceEnvName = "LCM_ALIVE_INTERFACE_PATH";
 constexpr uint32_t kDefaultProcessExecutionError = 1U;
 
-uint32_t defaultProcessorAffinityMask() {
-    return static_cast<uint32_t>((1ULL << score::lcm::internal::osal::getNumCores()) - 1ULL);
+uint64_t defaultProcessorAffinityMask() {
+    return (1ULL << score::lcm::internal::osal::getNumCores()) - 1ULL;
 }
 }  // namespace
